+++ src/main/java/at/hannibal2/skyhanni/features/inventory/ItemDisplayOverlayFeatures.kt
@@ -100,19 +100,6 @@ class ItemDisplayOverlayFeatures {
         }
 
         if (SkyHanniMod.feature.inventory.itemNumberAsStackSize.contains(7)) {
-            if (itemName.contains("Wishing Compass")) {
-                for (line in item.getLore()) {
-                    wishingCompassPattern.matchMatcher(line) {
-                        val uses = group("amount")
-                        if (uses != "3") {
-                            return uses
-                        }
-                    }
-                }
-            }
-        }
-
-        if (SkyHanniMod.feature.inventory.itemNumberAsStackSize.contains(8)) {
             if (itemName.contains("Kuudra Key")) {
                 return when (itemName) {
                     "Kuudra Key" -> "§a1"
@@ -125,7 +112,7 @@ class ItemDisplayOverlayFeatures {
             }
         }
 
-        if (SkyHanniMod.feature.inventory.itemNumberAsStackSize.contains(9)) {
+        if (SkyHanniMod.feature.inventory.itemNumberAsStackSize.contains(8)) {
             if (InventoryUtils.openInventoryName() == "Your Skills") {
                 if (item.getLore().any { it.contains("Click to view!") }) {
                     if (CollectionAPI.isCollectionTier0(item.getLore())) return "0"
@@ -137,7 +124,7 @@ class ItemDisplayOverlayFeatures {
             }
         }
 
-        if (SkyHanniMod.feature.inventory.itemNumberAsStackSize.contains(10)) {
+        if (SkyHanniMod.feature.inventory.itemNumberAsStackSize.contains(9)) {
             if (InventoryUtils.openInventoryName().endsWith(" Collections")) {
                 val lore = item.getLore()
                 if (lore.any { it.contains("Click to view!") }) {
@@ -152,7 +139,7 @@ class ItemDisplayOverlayFeatures {
             }
         }
 
-        if (SkyHanniMod.feature.inventory.itemNumberAsStackSize.contains(11)) {
+        if (SkyHanniMod.feature.inventory.itemNumberAsStackSize.contains(10)) {
             if (itemName.contains("Rancher's Boots")) {
                 for (line in item.getLore()) {
                     rangerBootsSpeedCapPattern.matchMatcher(line) {
@@ -162,7 +149,7 @@ class ItemDisplayOverlayFeatures {
             }
         }
 
-        if (SkyHanniMod.feature.inventory.itemNumberAsStackSize.contains(12)) {
+        if (SkyHanniMod.feature.inventory.itemNumberAsStackSize.contains(11)) {
             if (item.getInternalName_old() == "LARVA_HOOK") {
                 for (line in item.getLore()) {
                     "§7§7You may harvest §6(?<amount>.).*".toPattern().matchMatcher(line) {
@@ -177,7 +164,7 @@ class ItemDisplayOverlayFeatures {
             }
         }
 
-        if (SkyHanniMod.feature.inventory.itemNumberAsStackSize.contains(13)) {
+        if (SkyHanniMod.feature.inventory.itemNumberAsStackSize.contains(12)) {
             if (item.getInternalName_old() == "POTION") {
                 item.name?.let {
                     "Dungeon (?<level>.*) Potion".toPattern().matchMatcher(it.removeColor()) {